// ffb/src/services/http/Types.ts
export interface iPessoaSnakeCase {
  id: number;
  nome: string;
  data_nasc: string;
  cpf: string;
  altura: number;
  peso: number;
  sexo: string;
}

export interface iPessoa {
  id: number;
  nome: string;
  dataNascimento: Date;
  cpf: string;
  altura: number;
  peso: number;
  sexo: string;
}